Split out query params methods for Works
